+++ package.json
@@ -15,11 +15,15 @@
   "type": "commonjs",
   "main": "index.js",
   "scripts": {
-    "start": "node src/index.js",
+    "start": "node server.js",
+    "dev": "nodemon server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
     "express": "^4.21.2",
     "jquery": "^3.7.1"
+  },
+  "devDependencies": {
+    "nodemon": "^3.1.9"
   }
 }
