+++ public/style.css
@@ -1,0 +1,3 @@
+:root {
+    background-color: #acacac;
+}
